+++ database/migrations/2021_09_19_180437_create_acept_reservations_table.php
@@ -23,7 +23,7 @@ class CreateAceptReservationsTable extends Migration
             $table->integer('capacity');
             $table->string('state');
             $table->string('comment');
-            $table->float('pay');
+            $table->integer('pay');
             $table->timestamps();
         });
     }
