+++ database/migrations/2021_09_22_180905_create_reservations_table.php
@@ -23,7 +23,7 @@ class CreateReservationsTable extends Migration
             $table->integer('capacity');
             $table->string('state');
             $table->string('comment');
-            $table->float('pay');
+            $table->integer('pay');
             $table->timestamps();
         });
     }
